+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Rainbow CSV for Visual Studio Code Change Log
 
+## 3.21.0
+* Use Virtual alignment by default for the Align button - 2nd rollout attempt.
+* Add `Copy for Markdown export` command.
+
 ## 3.20.0
 * Add `Copy for Excel / Sheets export` command.
 * RBQL: remove trailing spaces from fields by default.